pyproject.toml
[project]
name = "investmentadvisoragent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.11.7",
    "fmpsdk>=20250102.0",
    "google-genai>=1.24.0",
    "newsapi-python>=0.2.7",
    "chromadb>=1.0.15",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ruff>=0.12.2",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "SIM", # flake8-simplify
    "ANN", # flake8-annotations
    "D",   # flake8-docstrings

]

ignore = [
    "E501",
    "B905",
    "D213",
    "D203",
    "D205",
    "D100",
    "D104"
]

